requirements.txt
notebook
matplotlib
pandas
ipywidgets
pathlib
torch 
torchvision 
torchaudio
pytorch-lightning
black
tabulate
torch-tb-profiler
hydra-core
hydra_colorlog
dvc[all]
boto3
pipreqs
wandb
nose
torchserve
torch-model-archiver
SoundFile
python-crontab